Refactor: Period 관련 객체들 common 패키지로 분리
