+++ server/src/main/java/com/ryc/api/v2/common/infra/mapper/PeriodMapper.java
@@ -1,7 +1,7 @@
-package com.ryc.api.v2.announcement.infra.mapper;
+package com.ryc.api.v2.common.infra.mapper;
 
-import com.ryc.api.v2.announcement.domain.vo.Period;
-import com.ryc.api.v2.announcement.infra.vo.PeriodVO;
+import com.ryc.api.v2.common.domain.Period;
+import com.ryc.api.v2.common.infra.vo.PeriodVO;
 
 public class PeriodMapper {
