+++ server/src/main/java/com/ryc/api/v2/email/domain/event/InterviewReservationEmailEvent.java
@@ -1,6 +1,6 @@
 package com.ryc.api.v2.email.domain.event;
 
-import com.ryc.api.v2.announcement.domain.vo.Period;
+import com.ryc.api.v2.common.domain.Period;
 
 import lombok.Builder;
