+++ server/src/main/java/com/ryc/api/v2/interview/infra/entity/InterviewSlotEntity.java
@@ -5,8 +5,8 @@ import java.util.List;
 
 import jakarta.persistence.*;
 
-import com.ryc.api.v2.announcement.infra.vo.PeriodVO;
 import com.ryc.api.v2.common.infra.entity.BaseEntity;
+import com.ryc.api.v2.common.infra.vo.PeriodVO;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
